C Project- University Management System
This project is model demonstration of how some university or big institution’s portal should look like so that it is easier for students to be aware of their updated class and bus time-tables. 
Here students can have all the necessary information they will need. For each year student, there is a class timetable with respect to his discipline and section. Every student can check the bus schedule for required day. This project is important for fresh students who seem to have no clue about university schedules and get a hard time finding classrooms for respective courses. In addition to this, this project allows administrator that is teacher, to do any type of update in timetable without having to get in trouble with long procedures of editing recorded data. The two main users of this project are Administrator and Student, and two main divisions of this project are class schedule and bus routine.